reorganize Addon class packages
